todo

+++ cgi-bin/removeCourse.py
@@ -13,22 +13,35 @@ course_id = str(form["course_id"].value)
 
 db_connection = sqlite3.connect('curricularUnits.db')
 cursor = db_connection.cursor()
-
-try:
-    cursor.execute('DELETE FROM Courses WHERE course_id = ?', [course_id])
+cursor.execute("SELECT COUNT(*) FROM Courses WHERE course_id = ?;", [course_id])
+nC = cursor.fetchone()
+if(nC==0):
     print("Content-type:text/html\n\n \
-        <!DOCTYPE html> \
-        <head> \
-            <h3> Course " + course_id + " was removed </h3> \
-        </head> \
-        <body> ")
-except sqlite3.Error as er:
-	print("Content-type:text/html\n\n \
         <!DOCTYPE html> \
         <head> \
             <h3> Course " + course_id + " does not exist </h3> \
         </head> \
-        <body> ")
+        body> ")
+else: 
+    try:
+        cursor.execute('DELETE FROM Courses WHERE course_id = ?', [course_id])
+        print("Content-type:text/html\n\n \
+            <!DOCTYPE html> \
+            <head> \
+                <h3> Course " + course_id + " was removed </h3> \
+            </head> \
+            <body> ")
+    except sqlite3.Error as er:
+	    print("Content-type:text/html\n\n \
+            <!DOCTYPE html> \
+            <head> \
+                <h3> Course " + course_id + " does not exist </h3> \
+            </head> \
+            body> ")
+
+
+    
+ 
 db_connection.commit()
 db_connection.close()
